split map.go - only functions of embedded pool mode.
add hinting.
