+++ pool2.go
@@ -6,7 +6,6 @@ import (
 	"io"
 	"strings"
 	"sync"
-	"sync/atomic"
 	"unsafe"
 
 	"github.com/kazu/elist_head"
@@ -146,30 +145,6 @@ func (sp *samepleItemPool) _init(cap int) {
 	//sp.Init()
 }
 
-const (
-	getEmpty      byte = 1
-	getLargest         = 2
-	getNoCap           = 3
-	requireInsert      = 4
-)
-
-func (sp *samepleItemPool) state4get(reverse uint64, tail int) byte {
-
-	if len(sp.items) == 0 {
-		return getEmpty
-	}
-	if cap(sp.items) == len(sp.items) {
-		return getNoCap
-	}
-
-	last := &sp.items[tail]
-	if last.reverse < reverse {
-		return getLargest
-	}
-	return requireInsert
-
-}
-
 func (sp *samepleItemPool) validateItems() error {
 	old := -1
 	empty := elist_head.ListHead{}
@@ -210,218 +185,6 @@ func (sp *samepleItemPool) validateItems() error {
 		}
 	}
 	return nil
-
-}
-
-var lastgets []byte = nil
-
-func lazyUnlock(mu *sync.Mutex) {
-	if mu != nil {
-		mu.Unlock()
-	}
-}
-
-type unlocker func(mu *sync.Mutex)
-
-func (sp *samepleItemPool) appendLast(mu *sync.Mutex) (newItem MapItem, nPool *samepleItemPool, fn unlocker) {
-
-	if mu != nil {
-		mu.Lock()
-		fn = lazyUnlock
-	}
-
-	olen := len(sp.items)
-	sp.items = sp.items[:olen+1]
-	new := &sp.items[olen]
-	new.Init()
-	return new, nil, fn
-
-}
-
-// requireInsert
-
-func (sp *samepleItemPool) insertToPool(reverse uint64, mu *sync.Mutex) (newItem MapItem, nPool *samepleItemPool, fn unlocker) {
-	if mu != nil {
-		mu.Lock()
-		fn = lazyUnlock
-	}
-
-	olen := len(sp.items)
-	ocap := cap(sp.items)
-
-	// require insert
-	//FIXME: should bsearch
-	if IsDebug() {
-		var b strings.Builder
-		for i := range sp.items {
-			sp.items[i].PtrMapHead().dump(&b)
-		}
-		Log(LogDebug, "B: itemPool.items\n%s\n", b.String())
-	}
-	// FIXME: should disable not IsDebug()
-	//sp.validateItems()
-	if olen != len(sp.items) {
-		Log(LogDebug, "update olen")
-		return sp.getWithFn(reverse, nil)
-	}
-	//olen = len(sp.items)
-	nlen := int64(olen)
-
-	for i := 0; i < olen; i++ {
-		//for i := range sp.items {
-		if sp.items[i].IsIgnored() {
-			continue
-		}
-		if sp.items[i].reverse < reverse {
-			continue
-		}
-		if i == olen-1 {
-			//fmt.Printf("invalid")
-		}
-		if !atomic.CompareAndSwapInt64(&nlen, int64(len(sp.items)), nlen+1) {
-			Log(LogDebug, "update olen")
-			return sp.getWithFn(reverse, nil)
-		}
-		var err error
-		prevItem := sp.items[0].ListHead.Prev()
-		nextItem := sp.items[olen-1].ListHead.Next()
-
-		// copy to new slice
-		newItems := make([]SampleItem, olen+1, maxInts(ocap, olen+1))
-		if i > 0 {
-			copy(newItems[0:i], sp.items[0:i])
-		}
-		copy(newItems[i+1:], sp.items[i:])
-
-		newListHead := &elist_head.ListHead{}
-		newListTail := &elist_head.ListHead{}
-		elist_head.InitAsEmpty(newListHead, newListTail)
-
-		middle := nextListHeadOfSampleItem()
-		for i := 0; i < olen+1; i++ {
-			newItems[i].ListHead = middle
-		}
-
-		err = newListHead.ReplaceNext(&newItems[0].ListHead, &newItems[olen].ListHead, newListTail)
-		if err != nil {
-			Log(LogFatal, "replace fail")
-		}
-		first := EmptyMapHead.fromListHead(newListHead.Next())
-		last := EmptyMapHead.fromListHead(newListTail.Prev())
-		pOpts := elist_head.SharedTrav(list_head.WaitNoM())
-		newItems[i].MarkForDelete()
-		elist_head.SharedTrav(pOpts...)
-		newItems[i].Init()
-
-		// for debug
-		first = EmptyMapHead.fromListHead(newListHead.Next())
-		last = EmptyMapHead.fromListHead(newListTail.Next())
-		_, _ = first, last
-
-		err = prevItem.ReplaceNext(newListHead.Next(), newListTail.Prev(), nextItem)
-		if err != nil {
-			Log(LogFatal, "fail to replace newItems")
-		}
-		oldItems := sp.items
-		sp.items = newItems
-
-		// for debug
-		oldItemFirst := oldItems[0].Prev().Next().PtrMapHead()
-		oldItemNext := oldItems[olen-1].Next().PtrMapHead()
-		_ = oldItemFirst
-		_ = oldItemNext
-		ItemNext := sp.items[olen].Next().PtrMapHead()
-		_ = ItemNext
-
-		if IsDebug() {
-			var b strings.Builder
-			for i := range sp.items {
-				sp.items[i].PtrMapHead().dump(&b)
-			}
-			fmt.Printf("A: itemPool.items\n%s\n", b.String())
-		}
-
-		outside := sp.items[olen].Next()
-		_ = outside
-		if olen != i && olen-1 != i && olen-1 > 0 && sp.items[olen-1].PtrListHead().Next() != sp.items[olen].PtrListHead() {
-			toNext := sp.items[olen-1].PtrListHead().Next()
-			next := sp.items[olen].PtrListHead()
-			Log(LogFatal, "not connect sp.items[olen-1]=%p -> sp.items[olen]=%p ", toNext, next)
-		}
-		if olen != i && olen-1 != i && olen-1 > 0 && sp.items[olen].PtrListHead().Prev() != sp.items[olen-1].PtrListHead() {
-			c := sp.items[olen].PtrListHead().Prev()
-			p := sp.items[olen-1].PtrListHead()
-			Log(LogFatal, "not connect sp.items[olen-1]=%p <- sp.items[olen]=%p", p, c)
-		}
-
-		return &sp.items[i], nil, lazyUnlock
-	}
-	return sp.getWithFn(reverse, nil)
-	//return nil, nil, nil
-
-}
-func (sp *samepleItemPool) getWithFn(reverse uint64, mu *sync.Mutex) (new MapItem, nPool *samepleItemPool, fn unlocker) {
-
-	lastActiveIdx := -1
-	olen := len(sp.items)
-
-	for i := olen - 1; i >= 0; i-- {
-		if sp.items[i].IsIgnored() {
-			continue
-		}
-		lastActiveIdx = i
-		break
-	}
-
-	defer func() {
-		if new == nil {
-			Log(LogWarn, "getWithFn(): item is nil")
-		}
-
-	}()
-
-	// for debug
-	//lastgets = append(lastgets, sp.state4get(reverse, lastActiveIdx))
-	switch sp.state4get(reverse, lastActiveIdx) {
-	case getEmpty, getLargest:
-		return sp.appendLast(mu)
-	case getNoCap:
-		fn, err := sp.expand(mu)
-		if err != nil {
-			Log(LogWarn, "pool.expand() require retry")
-		}
-		new, nPool, _ = sp.getWithFn(reverse, nil)
-		return new, nPool, fn
-	}
-
-	return sp.insertToPool(reverse, mu)
-
-}
-
-func nextListHeadOfSampleItem() elist_head.ListHead {
-
-	list := elist_head.NewEmptyList()
-
-	items := make([]SampleItem, 3)
-
-	list.Tail().InsertBefore(items[2].PtrListHead())
-	items[2].InsertBefore(items[1].PtrListHead())
-	items[1].InsertBefore(items[0].PtrListHead())
-
-	return items[1].ListHead
-}
-
-func (sp *samepleItemPool) findIdx(reverse uint64) (int, error) {
-
-	for i := range sp.items {
-		if sp.items[i].IsIgnored() {
-			continue
-		}
-		if reverse <= sp.items[i].reverse {
-			return i, nil
-		}
-	}
-	return -1, ErrIdxOverflow
 
 }
 
@@ -496,106 +259,6 @@ func (sp *samepleItemPool) DumpExpandInfo(w io.Writer, outers []unsafe.Pointer, 
 		mhead = EmptyMapHead.FromListHead(nCur).(*MapHead)
 		mhead.dump(w)
 	}
-
-}
-
-func (sp *samepleItemPool) split(idx int) (nPool *samepleItemPool, err error) {
-	return sp._split(idx, true)
-
-}
-func (sp *samepleItemPool) _split(idx int, connect bool) (nPool *samepleItemPool, err error) {
-
-	nlen := int64(len(sp.items))
-	if int(nlen) <= idx {
-		return nil, ErrIdxOverflow
-	}
-
-	nPool = &samepleItemPool{}
-
-	if !atomic.CompareAndSwapInt64(&nlen, int64(len(sp.items)), nlen+1) {
-		return sp._split(idx, connect)
-	}
-	nPool.items = sp.items[idx:]
-	sp.items = sp.items[:idx:idx]
-	nPool.Init()
-	//sp.validateItems()
-	//nPool.validateItems()
-
-	if connect && sp.PtrListHead().Next() != nil && sp.PtrListHead().Next() != sp.PtrListHead() {
-		_, err = sp.PtrListHead().Next().InsertBefore(nPool.PtrListHead())
-	}
-	return
-
-}
-
-func (sp *samepleItemPool) expand(mu *sync.Mutex) (unlocker, error) {
-	var fn unlocker
-	if mu != nil {
-		mu.Lock()
-		fn = lazyUnlock
-	}
-
-	olen := len(sp.items)
-	//ocap := cap(sp.items)
-
-	if olen != len(sp.items) {
-		Log(LogDebug, "update olen")
-		return sp.expand(nil)
-	}
-	nlen := int64(olen)
-
-	if !atomic.CompareAndSwapInt64(&nlen, int64(len(sp.items)), nlen+1) {
-		Log(LogDebug, "update olen")
-		return sp.expand(nil)
-	}
-
-	var err error
-	prevItem := sp.items[0].ListHead.Prev()
-	nextItem := sp.items[olen-1].ListHead.Next()
-
-	nCap := 0
-	if len(sp.items) < 128 {
-		nCap = 256
-	} else {
-		nCap = calcCap(len(sp.items))
-	}
-
-	newItems := make([]SampleItem, olen, nCap)
-	copy(newItems, sp.items[0:olen])
-
-	newListHead := &elist_head.ListHead{}
-	newListTail := &elist_head.ListHead{}
-	elist_head.InitAsEmpty(newListHead, newListTail)
-
-	err = newListHead.ReplaceNext(&newItems[0].ListHead, &newItems[olen-1].ListHead, newListTail)
-	if err != nil {
-		Log(LogFatal, "replace fail")
-	}
-
-	err = prevItem.ReplaceNext(newListHead.Next(), newListTail.Prev(), nextItem)
-	if err != nil {
-		Log(LogFatal, "fail to replace newItems")
-	}
-	oldItems := sp.items
-	sp.items = newItems
-
-	// for debug
-	oldItemFirst := oldItems[0].Prev().Next().PtrMapHead()
-	oldItemNext := oldItems[olen-1].Next().PtrMapHead()
-	_ = oldItemFirst
-	_ = oldItemNext
-	ItemNext := sp.items[olen-1].Next().PtrMapHead()
-	_ = ItemNext
-
-	if IsDebug() {
-		var b strings.Builder
-		for i := range sp.items {
-			sp.items[i].PtrMapHead().dump(&b)
-		}
-		fmt.Printf("A: itemPool.items\n%s\n", b.String())
-	}
-
-	return fn, nil
 
 }
